backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
PyJWT==2.12.1
Werkzeug==3.0.1
